First commit (warning: has errors)
'Meso-online' goals = ones that don't correspond to a site, but that we can track through their browser activity, such as their interactions with a nudge. For example, gratitude journaling ('write one thing that you're grateful for') and meditation ('do nothing for two minutes').

- Filters out meso-online goals from action-dropdown's choices (for now, until we associate an action script with each meso-online goal)
- Started making gratitude-prompt and gratitude-interstitial-screen components, a keep_gratefulness_journal goal, and an interstitial_gratitude_prompt intervention.

Has errors in the build, but committing work in progress to get back to working on the Duolingo-oriented release.