Final commit of the summer: generalized goal-setting
Created a goal-target-input component that creates an appropriate menu input for any goal passed as an attribute and added the component to the top of the site view pages.

One small bug remains: the graphs on the site view pages don't change immediately when the targets are changed, only on page refresh.

Also added properties to the goals called goal_statement_to_fill_in and goal_statement_units (optional, to override the units from the target property, which are still used for graphing). These are used by get_goal_statement() in goal_utils to populate the goal statement in positive-site-trigger-v2 as expected.