Nudge is functional!
Called "Mission Objective", this nudge asks you what you're going there to do before entering a site, then puts a banner with it at the top.

Added it to the Facebook interventions (as a generic one), not yet to the others.

Made an icon as well.

Next step is to extend this to add a time limit, which you set upfront or not ("Bouncer" or "1Minute Assassin" variants).

One bug is you can't type certain characters into the input field, like k and l. Not sure why, will investigate.